+++ ChangeLog.md
@@ -12,9 +12,24 @@
   a = 1
   ```
 
+### Loss backend
+
+* Fixed internal error with assertions with intra-tensor dependencies e.g. `forall x . x ! 0 < x ! 1`.
+
+* A `Bool` `@parameter` is now treated as an opaque value rather than causing an internal error.
+
 ### Solver backend
 
 * Fixed internal error that occassionally happened with non-linear specifications.
+
+### Agda backend
+
+* Fixed error where Agda library didn't type-check due to malformed transpose definition.
+
+### Python bindings
+
+* Fixed the error classes being unraisable through a `contextlib.contextmanager` on Python 3.11 and
+  later, which replaced them with a `FrozenInstanceError`.
 
 ## v0.28.0